chore(filenames): rename files in entities/, widgets/, app-pages/
