+++ src/app-pages/excursions/ui/ExcursionsSection.tsx
@@ -1,8 +1,10 @@
 import { ExcursionsList } from '@/widgets/excursions-list';
+
 import { useExcursionFiltering } from '@/features/excursion-filtering';
+
 import { Container, HeadingHighlight, HeadingTitle, Section, SectionHeading } from '@/shared/ui';
 
-import { excursions } from '../model/constants';
+import { excursions } from '../model/excursions.mock';
 
 export const ExcursionsSection = () => {
   const { filteredExcursions } = useExcursionFiltering(excursions);
