+++ src/entities/excursion/model/excursion.types.ts
