+++ src/entities/photo/model/photo.types.ts
