+++ src/entities/review/ui/Review.tsx
@@ -7,7 +7,7 @@ import { Quotes } from '@/shared/assets';
 import { cn } from '@/shared/lib';
 import { Avatar, AvatarFallback, AvatarImage, StarRating } from '@/shared/ui';
 
-import { type Review as ReviewType } from '../model/types';
+import { type Review as ReviewType } from '../model/review.types';
 
 type ReviewProps = {
   className?: string;
