+++ src/entities/user/index.ts
@@ -1,1 +1,1 @@
-export { type User } from './model/types';
+export { type User } from './model/user.types';
