+++ src/widgets/excursions-list/lib/sort-excursions-by-date.ts
